+++ jars/orm/src/main/java/it/govpay/bd/model/TipoVersamentoDominio.java
@@ -1,5 +1,7 @@
 package it.govpay.bd.model;
 
+import java.math.BigDecimal;
+
 import org.openspcoop2.generic_project.exception.NotFoundException;
 import org.openspcoop2.generic_project.exception.ServiceException;
 
@@ -239,7 +241,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 		else 
 			return this.getAvvisaturaMailPromemoriaScadenzaTipoDefault();
 	}
-	public Integer getAvvisaturaMailPromemoriaScadenzaPreavviso() {
+	public BigDecimal getAvvisaturaMailPromemoriaScadenzaPreavviso() {
 		if(this.getAvvisaturaMailPromemoriaScadenzaPreavvisoCustom() != null)
 			return this.getAvvisaturaMailPromemoriaScadenzaPreavvisoCustom();
 		else 
@@ -366,7 +368,7 @@ public class TipoVersamentoDominio extends it.govpay.model.TipoVersamentoDominio
 		else 
 			return this.getAvvisaturaAppIoPromemoriaScadenzaTipoDefault();
 	}
-	public Integer getAvvisaturaAppIoPromemoriaScadenzaPreavviso() {
+	public BigDecimal getAvvisaturaAppIoPromemoriaScadenzaPreavviso() {
 		if(this.getAvvisaturaAppIoPromemoriaScadenzaPreavvisoCustom() != null)
 			return this.getAvvisaturaAppIoPromemoriaScadenzaPreavvisoCustom();
 		else
